Function Calculator
You will program a function calculator. Each function can be a function of one variable (ex: 3x^2-6)
or zero variables (ex: 57^4-10). The variable will always be "x". You can use the calculator to give the
value of the function at any point or to compute the derivative of the function.